requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
asyncpg
pillow              # Для обработки изображений
pandas              # Для работы с CSV и XLSX
python-docx         # Для работы с DOCX
fpdf                # Для создания PDF из текста
PyPDF2              # Для извлечения текста из PDF
openpyxl            # Для сохранения файлов в формате Excel
python-dotenv
httpx
pytest
pydantic-settings
aiofiles
pytest-asyncio
